% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aMED.R
\name{aMED_health_groups}
\alias{aMED_health_groups}
\title{aMED_health_groups}
\usage{
aMED_health_groups()
}
\value{
names list of healthy and less healthy aMED components for scoring
}
\description{
Define the healthy or less-healthy groups for the aMED diet score. 
Methodology as described by Fung et al. (2005)
}
\author{
Robbie Pope
}
\keyword{internal}
