% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aMED.R
\name{aMED_score_logic}
\alias{aMED_score_logic}
\title{aMED_score_logic}
\usage{
aMED_score_logic(
  row,
  medians_df,
  female_low = female_low,
  female_high = female_high,
  male_low = male_low,
  male_high = male_high
)
}
\arguments{
\item{row}{row of long_df to score based on population median}

\item{medians_df}{df of study population median intakes, output of get_aMED_medians}

\item{female_low}{Lower limit of healthy female alcohol range.}

\item{female_high}{Upper limit of healthy female alcohol range.}

\item{male_low}{Lower limit of healthy male alcohol range.}

\item{male_high}{Upper limit of healthy male alcohol range.}
}
\value{
score of 1 or 0 depending on aMED scoring logic.
}
\description{
Below median intake is scored 0 for positive foods whereas equal to or above 
intake is scored 1 for negative foods. For alcohol the score is sex specific 
and based on a range. Default or Females is a score of 1 if between 5-15g d-1, 
otherwise 0. For Males, 1 if between 5-25g d-1, otherwise 0. Range can be updated 
in the function parameters.
}
\author{
Robbie Pope
}
\keyword{internal}
