% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aMED.R
\name{aMED_standard_groupings}
\alias{aMED_standard_groupings}
\title{aMED_standard_groupings}
\usage{
aMED_standard_groupings()
}
\value{
named list of aMED food & nutrient components for TwinsUK EPIC FFQ
}
\description{
Define the nine aMED amalgamated food, beverage and nutrient components 
Methodology as described by Fung et al. (2005)
FFQ line item and nutrient groupings according to Asnicar et al. (2021).
}
\author{
Robbie Pope
}
\keyword{internal}
