% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aMED.R
\name{assign_aMED_health_groups}
\alias{assign_aMED_health_groups}
\title{assign_aMED_health_groups}
\usage{
assign_aMED_health_groups(df)
}
\arguments{
\item{df}{Dataframe of aMED score components. Output of get_aMED_components}
}
\value{
df in long format of aMED components and healthful label.
}
\description{
Based on the aMED components as described by Fung et al. (2005).
}
\author{
Robbie Pope
}
\keyword{internal}
