% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DASH.R
\name{assign_dash_health_groups}
\alias{assign_dash_health_groups}
\title{assign_dash_health_groups}
\usage{
assign_dash_health_groups(df)
}
\arguments{
\item{df}{Dataframe of DASH score components. Output of get_dash_components}
}
\value{
df in long format of DASH components and healthful label.
}
\description{
Based on the DASH components as described by Fung et al. (2008).
}
\author{
Robbie Pope
}
\keyword{internal}
