% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DASH.R
\name{assign_dash_scores}
\alias{assign_dash_scores}
\title{assign_pdi_scores}
\usage{
assign_dash_scores(long_df)
}
\arguments{
\item{long_df}{long df with assigned ranked percentiles of intakes. Output from assign_percentile.}
}
\value{
Long df with added column for DASH_score.
}
\description{
Assign DASH scores to each component based on percentile and group classification.
}
\author{
Robbie Pope
}
\keyword{internal}
