% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pdi.R
\name{assign_health_groups}
\alias{assign_health_groups}
\title{assign_health_groups}
\usage{
assign_health_groups(df)
}
\arguments{
\item{df}{Dataframe of PDI food groups created by get_food_bev_groups.}
}
\value{
df in long format of PDI food groups and healthful label.
}
\description{
Based on the PDI groupings assign a healthy plant, less healthy plant or animal 
based food label for each of the 18 food groups (Satija et al. (2017)).
}
\author{
Robbie Pope
}
\keyword{internal}
