% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pdi.R
\name{assign_pdi_scores}
\alias{assign_pdi_scores}
\title{assign_pdi_scores}
\usage{
assign_pdi_scores(long_df)
}
\arguments{
\item{long_df}{long df with assigned ranked percentiles of intakes. Output from assign_percentile.}
}
\value{
Long df with added columns: PDI_score, hPDI_score, uPDI_score with rank based percentile 
        score for each food group for PDI, hPDI and uPDI.
}
\description{
Assign PDI, hPDI, and uPDI scores to each food group observation based on
percentile and group classification.
}
\author{
Robbie Pope
}
\keyword{internal}
