% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pdi.R
\name{assign_percentile}
\alias{assign_percentile}
\title{assign_percentile}
\usage{
assign_percentile(long_df, perc = 0.2)
}
\arguments{
\item{long_df}{long df output of assign_health_groups (must contain FoodGroup and Estimated_Intake).}

\item{perc}{proportion of distribution for each bin, float between 0 and 1 (default 0.2).}
}
\value{
long df with an additional integer column "Percentile" (1 = lowest).
}
\description{
Split study population into percentiles of intake for each food group,
assigning the percentile number in a new column (lowest percentile = lowest intake).
}
\author{
Robbie Pope
}
\keyword{internal}
