% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DASH.R
\name{calculate_dash}
\alias{calculate_dash}
\title{calculate_dash}
\usage{
calculate_dash(df_dash_components, perc = 0.2)
}
\arguments{
\item{df_dash_components}{A df of estimated intakes for the 8 amalgamated food & nutrient components for the DASH score (wide format).}

\item{perc}{proportion of distribution for each bin, float between 0 and 1 (default 0.2 for quintiles).}
}
\value{
A df containing the calculated DASH scores and estimated intakes of individual DASH components 
        using TwinsUK Standard groupings. 
        \describe{
        \item{DASH}{Overall PLant-based Dietary Index}
        \item{DASH_Components}{Individual est. intakes for 8 DASH components}
        }
}
\description{
Compute the DASH diet score using TwinsUK specific parameters and components.
DASH computation as described by Fung et al. (2008).
The calculate_dash function is a wrapper to perform all necessary steps to compute
each derivation of the score.
}
\references{
Fung, T. et al. (2008)
            *Adherence to a DASH-Style Diet and Risk of Coronary Heart Disease 
            and Stroke in Women*
            Archives of Internal Medicine 168(7):713-720.
}
\author{
Robbie Pope
}
