% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pdi.R
\name{calculate_pdi}
\alias{calculate_pdi}
\title{calculate_pdi}
\usage{
calculate_pdi(df_food_groups, perc = 0.2)
}
\arguments{
\item{df_food_groups}{A df of estimated intakes for the 18 amalgamated food groups for the PDI score (wide format).}

\item{perc}{proportion of distribution for each bin, float between 0 and 1 (default 0.2).}
}
\value{
A df containing the calculated PDI, hPDI and uPDI scores:
        \describe{
        \item{PDI}{Overall PLant-based Dietary Index}
        \item{hPDI}{Healthful Plant-based Dietary Index}
        \item{uPDI}{Unhealthful Plant-based Dietary Index}
        }
}
\description{
Compute the Plant-based Dietary Index (PDI) and it's healthful (hPDI) or 
unhealthful (uPDI) derivations. PDI computation as described by Satija et al. (2017).
The calculate_pdi function is a wrapper to perform all necessary steps to compute
each derivation of the score.
}
\references{
Satija, A. et al. (2017) *Healthful and unhealthful plant-based diets and the risk of coronary heart disease in US adults.* J Am Coll Cardiol 70(4):411–422.
}
\author{
Robbie Pope
}
