% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DASH.R
\name{compute_twinsuk_dash}
\alias{compute_twinsuk_dash}
\title{compute_twinsuk_dash}
\usage{
compute_twinsuk_dash(
  food.intakes,
  nutr.intakes,
  merge_col_name_food.intakes = "FFQ_ID",
  merge_col_name_nutr.intakes = "FFQ_ID",
  perc = 0.2
)
}
\arguments{
\item{food.intakes}{Pre-processed TwinsUK estimated intake .csv file (grams per day).
the first column should be the unique identifier for each questionnaire.}

\item{nutr.intakes}{Pre-processed TwinsUK estimated nutrient intakes .csv file,
the first column should be the unique identifier for each questionnaire.}

\item{merge_col_name_food.intakes}{name of column to merge estimated food and 
nutrient intakes for food grams per day .csv file. Default is 'FFQ_ID'.}

\item{merge_col_name_nutr.intakes}{name of column to merge estimated food and 
nutrient intakes for nutrient .csv file. Default is 'FFQ_ID'.
the first column should be the unique identifier for each questionnaire.}

\item{perc}{proportion of distribution for each bin, float between 0 and 1 (default 0.2 for quintiles).}
}
\value{
A df containing the calculated DASH scores and estimated intakes of individual DASH components 
        using TwinsUK Standard groupings. 
        \describe{
        \item{DASH}{Overall PLant-based Dietary Index}
        \item{DASH_Components}{Individual est. intakes for 8 DASH components}
        }
}
\description{
The compute_twinsuk_dash function is a wrapper of calculate_dash to compute the 
DASH diet score using TwinsUK specific parameters and food groupings.
DASH computation as described by Fung et al. (2008).
TwinsUK specific FFQ line item & nutrient components as described by Pope et al. (2021).
}
\references{
Fung, T. et al. (2008) *Adherence to a DASH-Style Diet and Risk of Coronary Heart Disease and Stroke in Women* Archives of Internal Medicine 168(7):713-720.
            Pope et al. (2025) *Faecal metabolites as a readout of habitual diet capture dietary interactions with the gut microbiome* Nature Communications 16(1):10051.
}
\author{
Robbie Pope
}
