% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pdi.R
\name{compute_twinsuk_pdi}
\alias{compute_twinsuk_pdi}
\title{compute_twinsuk_pdi}
\usage{
compute_twinsuk_pdi(food.intakes, perc = 0.2)
}
\arguments{
\item{food.intakes}{Pre-processed TwinsUK estimated intake .csv file (grams per day).
the first column should be the unique identifier for each questionnaire.}

\item{perc}{proportion of distribution for each bin, float between 0 and 1 (default 0.2 for quintiles).}
}
\value{
A df containing the calculated PDI, hPDI and uPDI scores and estimated intakes 
        of the amalgamated food groups using TwinsUK standard groupings.
        \describe{
        \item{PDI}{Overall PLant-based Dietary Index}
        \item{hPDI}{Healthful Plant-based Dietary Index}
        \item{uPDI}{Unhealthful Plant-based Dietary Index}
        \item{FooGroups}{Individual est. intakes for 18 food groups}
        }
}
\description{
The compute_twinsuk_pdi function is a wrapper of calculate_pdi to compute the 
Plant-based Dietary Index (PDI) and it's healthful (hPDI) or unhealthful (uPDI) 
derivations using TwinsUK specific parameters and food groupings.
PDI computation as described by Satija et al. (2017).
TwinsUK specific FFQ line item groupings as described by Asnicar et al. (2021).
}
\references{
Satija, A. et al. (2017) *Healthful and unhealthful plant-based diets and the risk of  coronary heart disease in US adults.* J Am Coll Cardiol 70(4):411–422.
            Asnicar et al. (2021) *Microbiome connections with host metabolism and habitual diet from 1,098 deeply phenotyped individuals.* Nature Medicine 27(2):321-332.
}
\author{
Robbie Pope
}
