% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DASH.R
\name{dash_health_groups}
\alias{dash_health_groups}
\title{dash_health_groups}
\usage{
dash_health_groups()
}
\value{
named list of healthy and less healthy DASH components for scoring
}
\description{
Define the healthy or less-healthy groups for the DASH diet score. 
Methodology as described by Fung et al. (2008)
}
\author{
Robbie Pope
}
\keyword{internal}
