% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DASH.R
\name{dash_standard_groupings}
\alias{dash_standard_groupings}
\title{dash_standard_groupings}
\usage{
dash_standard_groupings()
}
\value{
named list of DASH food & nutrient components for TwinsUK EPIC FFQ
}
\description{
Define the nine DASH amalgamated food, beverage and nutrient components 
Methodology as described by Fung et al. (2008)
FFQ line item and nutrient groupings according to Pope et al. (2025).
}
\author{
Robbie Pope
}
\keyword{internal}
