% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pdi.R
\name{define_health_groups}
\alias{define_health_groups}
\title{define_health_groups}
\usage{
define_health_groups()
}
\value{
named list of healthy plant, less healthy plant and animal based food 
        groupings.
}
\description{
Define the healthy plant, less healthy plant and animal based food groups for 
the PDI, hPDI and uPDI calculations (Satija et al. (2017)).
}
\author{
Robbie Pope
}
\keyword{internal}
