% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aMED.R
\name{get_aMED_medians}
\alias{get_aMED_medians}
\title{get_aMED_medians}
\usage{
get_aMED_medians(long_df)
}
\arguments{
\item{long_df}{long df of etimated intakes of aMED components. Output from assign_aMED_health_groups}
}
\value{
medians_df, a df of median intakes from the long_df
}
\description{
Compute median intake for each component of the aMED score.
}
\author{
Robbie Pope
}
\keyword{internal}
