% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DASH.R
\name{get_dash_components}
\alias{get_dash_components}
\title{get_dash_components}
\usage{
get_dash_components(
  food.intakes,
  nutr.intakes,
  merge_col_name_food.intakes = "FFQ_ID",
  merge_col_name_nutr.intakes = "FFQ_ID",
  stnd_components = "Y",
  alt_components = NULL
)
}
\arguments{
\item{food.intakes}{Pre-processed TwinsUK estimated intake .csv file (grams per day).
the first column should be the unique identifier for each questionnaire.}

\item{nutr.intakes}{Pre-processed TwinsUK estimated nutrient intakes .csv file,
the first column should be the unique identifier for each questionnaire.}

\item{merge_col_name_food.intakes}{name of column to merge estimated food and 
nutrient intakes for food grams per day .csv file. Default is 'FFQ_ID'.}

\item{merge_col_name_nutr.intakes}{name of column to merge estimated food and 
nutrient intakes for nutrient .csv file. Default is 'FFQ_ID'.
the first column should be the unique identifier for each questionnaire.}

\item{stnd_components}{Character either "Y" (yes) or "N" (no) to determine if 
standard Pope et al. (2025) groupings for DASH components are used.}

\item{alt_components}{Named list of custom food and nutrient component definitions, 
used only when stnd_components = "N". Each list element name should 
be a one of the DASH components, and each value a character vector of column names.}
}
\value{
Data frame with summed daily intakes for each food, beverage or nutrient 
        component of the DASH diet score.
}
\description{
Take pre-prepared TwinsUK grams per day and nutrient estimate .csv files and compute 
the estimated daily intakes for amalgamated food, beverage and nutrient components of 
the DASH diet score. Standard groupings (Pope et al. (2025)), or custom groupings for
each component can be used.
}
\author{
Robbie Pope
}
