% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_energy_adjust.R
\name{residual_energy_adjust}
\alias{residual_energy_adjust}
\title{residual_energy_adjust}
\usage{
residual_energy_adjust(intakes, energy_col, start_col, end_col)
}
\arguments{
\item{intakes}{Pre-processed TwinsUK estimated food/nutrient intake .csv file 
(grams per day). The first column should be the unique identifier 
for each questionnaire.}

\item{energy_col}{Numeric index of the energy intake column used for adjustment.}

\item{start_col}{Numeric index of the first nutrient or food group column.}

\item{end_col}{Numeric index of the final nutrient or food group column.}
}
\value{
Data frame containing:
        \describe{
        \item{metadata}{Original metadata columns}
        \item{unadjusted variables}{Original nutrient/food group intakes}
        \item{adjusted variables}{Energy-adjusted variables with suffix '_adj'}
        }
}
\description{
Nutrient density, residual method. Take raw estimated daily nutrient or food 
group intakes and adjust based on energy intake. (Willet et al. 1997). Recommended
to adjust raw nutrien estiamtes or summarised food groups. To get summarised 
food groups use function get_food_bev_groups with stnd_groups argument == 'Y'.
Energy variables (e.g. Energy_kcal, Energy_kj) are not adjusted.
}
\references{
Willett et al. (1997) *Adjustment for total energy intake in epidemiologic studies.* Am J Clin Nutr 65(4):1220S-1228S.
}
\author{
Robbie Pope
}
