% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pdi.R
\name{score_logic}
\alias{score_logic}
\title{score_logic}
\usage{
score_logic(long_df, positive_groups, negative_groups, score_name)
}
\arguments{
\item{long_df}{long df with assigned ranked percentiles of intakes. Output from assign_percentile.}

\item{positive_groups}{vector of groups to be scored positively}

\item{negative_groups}{vector of groups to be scored negatively}

\item{score_name}{description}
}
\value{
long_df with additional percentile scoring in an additional column for the type of pdi score
}
\description{
Split study population into percentiles of intake for each food group,
assigning the percentile number in a new column (lowest percentile = lowest intake).
}
\author{
Robbie Pope
}
\keyword{internal}
