% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_bev_groups.R
\name{standard_groupings}
\alias{standard_groupings}
\title{standard_groupings}
\usage{
standard_groupings()
}
\value{
named list of standard food groupings based on TwinsUK EPIC FFQ
}
\description{
Define standard amalgamated food and beverage groups based on TwinsUK EPIC 
FFQ line items (Asnicar et al. (2021)).
}
\author{
Robbie Pope
}
\keyword{internal}
