% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aMED.R
\name{sum_aMED_scores}
\alias{sum_aMED_scores}
\title{sum_aMED_scores}
\usage{
sum_aMED_scores(long_df)
}
\arguments{
\item{long_df}{long df with scored components. Output from assign_aMED_scores.}
}
\value{
df with summed aMED score for each individual FFQ and estimated intake for each component.
}
\description{
Sum the median based score for each aMED component to give the final aMED score
}
\author{
Robbie Pope
}
\keyword{internal}
