% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DASH.R
\name{sum_dash_scores}
\alias{sum_dash_scores}
\title{sum_dash_scores}
\usage{
sum_dash_scores(long_df)
}
\arguments{
\item{long_df}{long df with assigned ranked percentile scores. Output from assign_dash_scores.}
}
\value{
df with summed DASH score for each individual FFQ and estimated intake for each component.
}
\description{
Sum the ranked percentile scores for DASH score
}
\author{
Robbie Pope
}
\keyword{internal}
