% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_bev_groups.R
\name{sum_food_groups}
\alias{sum_food_groups}
\title{sum_food_groups}
\usage{
sum_food_groups(food.intakes, named_list)
}
\arguments{
\item{food.intakes}{Pre-processed TwinsUK estimated intake .csv file (grams per day).
the first column should be the unique identifier for each questionnaire.}

\item{named_list}{Named list of food and beverage group definitions, 
Each list element name should be a food group, and each value a 
character vector of column names.}
}
\value{
A list of numeric vectors (each of length equal to `nrow(data)`), where
        each vector represents the summed intake for a food or beverage group.
}
\description{
Define standard amalgamated food and beverage groups based on TwinsUK EPIC 
FFQ line items (Asnicar et al. (2021)).
}
\author{
Robbie Pope
}
\keyword{internal}
