% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pdi.R
\name{sum_pdi_scores}
\alias{sum_pdi_scores}
\title{sum_pdi_scores}
\usage{
sum_pdi_scores(long_df)
}
\arguments{
\item{long_df}{long df with assigned ranked percentile scores. Output from assign_pdi_scores.}
}
\value{
df with summed PDI, hPDI and uPDI scores for each indivdiual.
}
\description{
Sum the ranked percentile scores for each derivation of the PDI.
}
\author{
Robbie Pope
}
\keyword{internal}
