% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aMED.R
\name{assign_aMED_scores}
\alias{assign_aMED_scores}
\title{assign_aMED_scores}
\usage{
assign_aMED_scores(
  long_df,
  est.intakes,
  id_col = "FFQ_ID",
  sex_col = "Sex",
  female_low = 5,
  female_high = 15,
  male_low = 5,
  male_high = 25
)
}
\arguments{
\item{long_df}{long df of etimated intakes of aMED components. Output from assign_aMED_health_groups}

\item{est.intakes}{Pre-processed TwinsUK estimated intake .csv file (grams per day).
the first column should be the unique identifier for each questionnaire.}

\item{id_col}{Name of the column containing the unique identifier (default: 'FFQ_ID').}

\item{sex_col}{Name of the column containing sex information (default: 'Sex').}

\item{female_low}{Lower limit of female alcohol range.}

\item{female_high}{Upper limit of female alcohol range.}

\item{male_low}{Lower limit of male alcohol range.}

\item{male_high}{Upper limit of male alcohol range.}
}
\value{
Long df with added column for aMED_score.
}
\description{
Assign aMED scores to each component based on if intake is above or below study
population median or within specified alcohol range.
}
\author{
Robbie Pope
}
\keyword{internal}
