% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aMED.R
\name{sum_aMED_components}
\alias{sum_aMED_components}
\title{sum_aMED_components}
\usage{
sum_aMED_components(est.intakes, nutr.intakes, named_list)
}
\arguments{
\item{est.intakes}{Pre-processed TwinsUK estimated intake .csv file (grams per day).
the first column should be the unique identifier for each questionnaire.}

\item{nutr.intakes}{Pre-processed TwinsUK estimated nutrient intakes .csv file,
the first column should be the unique identifier for each questionnaire.}

\item{named_list}{Named list of food and nutrient group components, 
Each list element name should be a food/nutrient component of the aMED diet, 
and each value a character vector of column names.}
}
\value{
A list of numeric vectors (each of length equal to `nrow(data)`), where
        each vector represents the summed intake for a food or nutrient component.
}
\description{
Define standard amalgamated food and nutrient components of aMED diet score 
for TwinsUK EPIC FFQ
FFQ line items (Asnicar et al. (2021)).
}
\author{
Robbie Pope
}
\keyword{internal}
