% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DASH.R
\name{sum_dash_components}
\alias{sum_dash_components}
\title{sum_dash_components}
\usage{
sum_dash_components(
  est.intakes,
  nutr.intakes,
  named_list,
  merge_col_name_est.intakes = "FFQ_ID",
  merge_col_name_nutr.intakes = "FFQ_ID"
)
}
\arguments{
\item{est.intakes}{Pre-processed TwinsUK estimated intake .csv file (grams per day).
the first column should be the unique identifier for each questionnaire.}

\item{nutr.intakes}{Pre-processed TwinsUK estimated nutrient intakes .csv file,
the first column should be the unique identifier for each questionnaire.}

\item{named_list}{Named list of food and nutrient group components, 
Each list element name should be a food/nutrient component of the DASH diet, 
and each value a character vector of column names.}

\item{merge_col_name_est.intakes}{name of column to merge estimated food and 
nutrient intakes for food grams per day .csv file. Default is 'FFQ_ID'.}

\item{merge_col_name_nutr.intakes}{name of column to merge estimated food and 
nutrient intakes for nutrient .csv file. Default is 'FFQ_ID'.}
}
\value{
A list of numeric vectors (each of length equal to `nrow(data)`), where
        each vector represents the summed intake for a food or nutrient component.
}
\description{
Define standard amalgamated food and nutrient components of DASH diet score 
for TwinsUK EPIC FFQ
FFQ line items (Pope et al. (2025)).
}
\author{
Robbie Pope
}
\keyword{internal}
